move utility functions to testing
